backend/python_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.2.2
transformers>=4.36.0
Pillow==10.1.0
numpy==1.26.2
pydantic==2.7.0
python-multipart==0.0.6
openai==1.3.5
torchaudio==2.2.2
pydantic-settings==2.5.2
torchvision==0.17.2
open_clip_torch
ftfy
pinecone-client==4.1.2
pinecone-plugin-inference==3.1.0
python-dotenv
python-dateutil
psutil